+++ CHANGELOG.md
@@ -56,6 +56,7 @@ at anytime.
   * Added abandon information (claim name, id, address, amount, balance_delta and nout) about claims, supports, and updates to `transaction_list` results under `abandon_info` key
   * Added `permanent_url` attribute to `channel_list_mine`, `claim_list`, `claim_show`, `resolve` and `resolve_name` API calls through lbryio/lbryum#203
   *
+  * Added `--conf` CLI flag to lbrynet-cli tool to specify an alternative config file
 
 ### Changed
   * claim_show API command no longer takes name as argument